requirements.txt
Django>=3.0,<4.0
tensorflow
numpy
djangorestframework